refactor: extract token file helpers to shared layer
Move getTokenFilePath() and readTokenFromFile() from src/server/auth/token-store.ts
to src/shared/auth/token-file.ts, eliminating the CLI→server layer violation.
token-store.ts re-exports both for backward compat; its internal functions
(writeTokenToFile, loadOrCreateToken) now import from shared. Update
rotate-token.ts to import directly from shared. Update test mock path to match.
